+++ application/profile/views.py
@@ -5,7 +5,7 @@ from contextlib import suppress
 from application.fhir.connect import smart
 from application.fhir.search import ResourceFinder
 from application.profile.forms import EditProfileForm
-from application.oauth.util import oauth_required
+from application.oauth.utils import oauth_required
 
 profile_bp = Blueprint(
     'profile_bp',
